Move address repository to libraries/shared
